+++ kompendium-core/src/main/kotlin/org/leafygreens/kompendium/annotations/HeaderParam.kt
@@ -1,5 +1,0 @@
-package org.leafygreens.kompendium.annotations
-
-@Retention(AnnotationRetention.RUNTIME)
-@Target(AnnotationTarget.PROPERTY)
-annotation class HeaderParam(val description: String = "")